threeC/requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
fancycompleter==0.9.1
google-api-core==1.17.0
google-api-python-client==1.8.3
google-auth==1.15.0
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
googleapis-common-protos==1.51.0
httplib2==0.18.1
idna==2.9
multidict==4.7.6
oauth2client==4.1.3
oauthlib==3.1.0
pdbpp==0.10.2
protobuf==3.12.1
py3cw==0.0.16
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.6.1
pygsheets==2.0.3.1
pyrepl==0.9.0
pytz==2020.1
PyYAML==5.3.1
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
six==1.15.0
slackclient==2.6.0
uritemplate==3.0.1
urllib3==1.25.9
wmctrl==0.3
yarl==1.4.2
